refactor: Change Session.Id to ulong to match SessionId in Cart and User

+++ Ecommerce.Entity/Session.cs
@@ -2,7 +2,7 @@
 
 public class Session
 {
-    public uint Id { get; set; }
+    public ulong Id { get; set; }
     public uint? CartId { get; set; }
     public Cart? Cart { get; set; }
     public uint? UserId {get;set;}
